html/syntax/parsing/the-input-byte-stream/the-input-byte-stream-026.html

<?xml version="1.0" encoding="ISO-8859-15"?><!DOCTYPE html>
<html  lang="en" >
<head>
 <meta charset="ISO-8859-1" > <title>XML declaration vs meta charset</title>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<link rel='help' href='http://www.w3.org/TR/html5/syntax.html#the-input-byte-stream'>
<script src="../../../../resources/testharness.js"></script>
<script src="../../../../resources/testharnessreport.js"></script>
<link rel="stylesheet" type="text/css" href="../../../../resources/testharness.css">
<meta name="assert" content="[Exploratory] The XML declaration has a lower precedence than an encoding declaration in a meta charset attribute.">
<meta name='flags' content='http'>
<style type='text/css'>
.test div { width: 50px; }.test div { width: 90px; }
</style>
<link rel="stylesheet" type="text/css" href="support/encodingtests-1.css">
</head>
<body>



<div class='test'><div id='box' class='ýäè'>&#xA0;</div></div>


<!--Notes:

The page contains an encoding declaration in a meta charset attribute that attempts to set the character encoding to ISO 8859-1, but the file starts with an XML declaration that tries to set the encoding to ISO 8859-15.

The test contains a div with a class name that contains the following sequence of bytes: 0xC3 0xBD 0xC3 0xA4 0xC3 0xA8. These represent different sequences of characters in ISO 8859-15, ISO 8859-1 and UTF-8. The external, UTF-8-encoded stylesheet contains a selector <code>.test div.&#x00C3;&#x00BD;&#x00C3;&#x00A4;&#x00C3;&#x00A8;</code>. This matches the sequence of bytes above when they are interpreted as ISO 8859-1. If the class name matches the selector then the test will pass.

-->
<script>
test(function() {
assert_equals(document.getElementById('box').offsetWidth, 100);
}, " ");
</script>

<div id='log'></div>

</body>
</html>
